ci: add static analysis workflow (clang-tidy, cppcheck) and fix codebase findings
- Add .clang-tidy configuration and quality.yml CI workflow over GUI and headless builds
- Add static-analysis.sh and cppcheck-compile-db.py to run clang-tidy and cppcheck via bear
- Resolve performance, modern C++, and bugprone findings across src/ (const refs, override, casts)
- Suppress vetted false positives with clear inline rationale

+++ src/vpn/OpenVpnConfigPolicy.h
@@ -72,7 +72,7 @@ struct Verdict
 
 //! Largest profile accepted. A real one is a few kilobytes; the cap bounds the
 //! work done on a file that root is about to read.
-constexpr qint64 kMaxConfigBytes = 1024 * 1024;
+constexpr qint64 kMaxConfigBytes = qint64(1024) * 1024;
 
 //! Inspect the bytes of a .ovpn profile.
 Verdict inspect(QByteArray const &config);
